requirements.txt
asgiref==3.6.0
autopep8==2.0.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
chatbot==1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
persiantools==3.0.1
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
